feat: 재입력 로직 구현, 책임 분리
-게임 결과 산출 책임을 GameStatus에게 위임
-게임 턴 관리 책임을 ChessGame에게 위임
